Fix TypeScript description
The prior description talks about TypeScript being a "simplification" of JavaScript, which is an odd thing to
to say about TypeScript. And it fails to mention that TypeScript adds a type system on top of JavaScript, which
is the most salient aspect of TypeScript. This replaces the short description with the subheading from
https://www.typescriptlang.org/, and replaces part of the long description with something more accurate.