Add innogames/igcommit
This is a Git pre-receive hook to check commits and code style.  The tool
automatically detects most of the items on this page and executes them
accordingly on changed files.  Unlike the "pre-commit" project listed here,
pre-receive hooks run on the Git server side.  This has advantages and
disadvantages that are listed on the project's README.