Update styleguide.md
I made the following  changes for your consideration:

First, the grammar stuff:
-Updated periods to appear inside of quotes throughout (example: Never abbreviate "pull request".  to "Never abbreviate "pull request.")
-Corrected "it's" usage under the Words that can be tricky / Open source section

And a few small style changes throughout.

I really like this style guide.  It lays out the information in a nice, concise manner. It's certainly helpful to me as a new GitHub user!

Best and hope my suggested changes are helpful --

--Alissa