Index.md
So you want to join Anonymous? -
You can not join Anonymous. Anonymous is not an organization. It is not a club, a party or even a movement. There is no charter, no manifest, no membership fees.

All we are is people who travel a short distance together -- much like commuters who meet in a bus or tram: For a brief period of time we have the same route, share a common goal, purpose or dislike. And on this journey together, we may well change the world.

Nobody could say: you are in, or you are out. Do you still want to join Anonymous? Well, you are in if you want to.

How to get in contact with others? -
Anonymous has no centralized infrastructure. We use existing facilities of the Internet, especially social networks, and we are ready to hop on to the next one if this one seems compromised, is under attack, or starts to bore us. Look for terms like "anonymous", "anonops" and other keywords that might be connected to our activities.

How do I recognize other Anonymous? -
We come from all places of society: We are students, workers, clerks, unemployed; We are young or old, we wear smart clothes or rugs, we are hedonists, ascetics, joy riders or activists. We come from all races, countries and ethnicities. We are many.
We are your neighbours, your co-workers, your hairdressers, your bus drivers and your network administrators. We are the guy on the street with the suitcase and the girl in the bar you are trying to chat up. We are anonymous. Many of us like to wear Guy Fawkes masks on demonstrations. Some of us even show them in their profile pictures in social networks. That helps to recognize each other.

Have you been infiltrated? -
If you talk to another Anonymous, you will never know who he is. He may be a hacker, cracker, phisher, agent, spy, provoker -- or just the guy from next door. Or his daughter. It is not illegal to be Anonymous. Nor is it illegal to wear Guy Fawkes masks. Keep that in mind. If you personally have not been involved in illegal activities, you have nothing to worry, no matter whom you talk to; If you have, it is wise not to talk about it. To no one.

How do I protect my privacy? -
Invent an alias, a nick, a pseudonym ... call it as you will, just invent something. Then register a mail account in that name with one of the big mail providers. Use this email address to register your Twitter, Facebook, etc. accounts. Make sure to clear all cookies before you start using your new identity, or better use a different web browser for Anonymous than for your other activities.

How many Anonymous are there? -
We are more than you think. We are more than anybody thinks. We are many.
And you are now one of us. 
Welcome to Anonymous.

We are Anonymous.
We are Legion.
We do not forgive.
We do not forget.
Expect us.

- Watch some of our other videos -
Anonymous Documentary - How Anonymous Hackers Changed the World Full Documentary: