Update index.md
David Chaum came up with the concept of Mix networks in 1981, Stuart Haber and W. Scott Stornetta came up with cryptographic and secured blocks in 1991. Satoshi coined the term blockchain in 2008 prior to releasing Bitcoin in 2009. I think it is fair that not only Satoshi is credited as the sole inventor in the Github explore page, and you change it to, "David Chaum, Stuart Haber, W. Scott Stornetta, Satoshi Nakamoto and various authors", which is technically more historically accurate.