Update topics/common-lisp/index.md
The description had newlines which resulted in two newlines when
rendering. Instead we the entire text of the description is included
in a single line, as it is one paragraph. Also markdown links are not
rendered, remove them.

Add the language related topic. It seems to be pattern. cf. ruby,
python and clojure.