Create Risc-v CPU/Cores Collection
The market for processors and microcontrollers is increasingly heated,
and with the launch of the ISA (Instruction Set Achitecture) RISC-V, an
open specification, it opens up a new opportunity for those who want to
act either by researching or collaborating with new processors and
microcontrollers.

Those who master the synthesis of Hardware with FPGA, can also propose
more concretely new approaches for microcontrollers taking advantage of
codes already written for the new architecture.

In this collection I try to present some renowned Colors, and open the
opportunity for other colleagues to collaborate with their suggestions.