Initial commit of index.md for cscw topic
I am adding this because I think it will be helpful to at least
expand the abbreviation CSCW, which stands for Computer-supported
cooperative work.  A survey of the repositories to which the
topic has already been applied reveals that they are in harmony
with the interpretation of CSCW as Computer-supported cooperative
work, and some of them expand the term in their README.md files.