Softened language around conflicts of interest
The goal behind this policy is to prevent people from advertising
a product or a pet project, but the wording restricted _anyone_
with a maintainer/contributor interest in a project from changing
the topic, which was too restrictive. This change makes the wording
more closely aligned with the actual policy.