Rename API.md to API.md
##  Making edits to a topic

If you're making edits to a topic page, the following fields are available for use. Not all fields are required.

### aliases
(if applicable) Synonyms for the topic name. For example, `react` and `reactjs` are aliases. You
should only list an alias for your topic if the majority of repositories using that alias are
referring to the same subject matter as the topic. You should not list another topic as an alias
if the alias is a superset of your topic.

For example, `api` makes sense in the `related` field for the `graphql` topic, but because many
repositories tagged with `api` are _not_ be associated with `graphql`, `api` does not make sense
in the `aliases` field for `graphql`.

Each alias must be formatted like that topic's `topic` field (same as the URL slug). Acceptable formatting:

* Starts with a letter or number
* Contains only letters, numbers, and hyphens
* At most 35 characters long

### created_by
(if applicable) Names of the people and/or organizations who authored the topic (NOT your name). For example, `Jordan Walke` is the author of `react`.

### display_name
(required) The topic name that will be displayed on the topic page (ex. `React`). Should use proper noun capitalization. Emoji are not allowed.

### github_url
(if applicable) URL of the topic's official GitHub organization or repository. Must start with `https://github.com/`.

### logo
(if applicable) The official logo associated with that topic. You must have permission to use this logo. If no official logo exists, do not include an image.

If you're submitting content for a topic page, upload the image to the topic's folder and put its name (ex. `logo.png`) here. The image must be square, `*.png` format, dimensions 288x288 and no larger than 75 kB. The file name must be the same as the `topic` with an image extension.

### related
(if applicable) Any related topics you can think of. Related topic suggestions are automatically generated by GitHub, but you have the option to call out any specific topics here.

Each related topic must be formatted like that topic's `topic` field (same as the URL slug). Acceptable formatting:

* Starts with a letter or number
* Contains only letters, numbers, and hyphens
* At most 35 characters long

Formatted as `topic1, topic2, topic3`.

### released
(if applicable) Date of first release. Formatted as `MONTH DD, YYYY`, `MONTH YYYY`, or just `YYYY`.

### short_description
(required) A short description of the topic, which will be used on the Explore homepage, Topics subpage, and other preview areas. Must be 130 characters or less. Emoji are not allowed.

### topic
(required) Name of the topic, which will be used in the URL, e.g. `https://github.com/topics/[URL]`. Formatted as lowercase.

Acceptable formatting:

* Starts with a letter or number
* Contains only letters, numbers, and hyphens
* At most 35 characters long

### url
(if applicable) URL to the topic's official website

### wikipedia_url
(if applicable) URL to a Wikipedia article about the topic

### The body of your document
(required) A longer description of the topic, which will appear on its topic page. Must be 1,000 characters or less. Should not be the same as `short_description`. Some Markdown is allowed, such as links. Emoji are allowed.

##  Making edits to a collection

If you're making edits to a collection, the following fields are available for use. Not all fields are required.

### items
A [YAML list](https://en.wikipedia.org/wiki/YAML#Basic_components) containing any of the following values:

- GitHub repository path (e.g. `defunkt/dotjs`)
- GitHub username (e.g. `defunkt`)
- GitHub organization (e.g. `github`)
- Any web URL (e.g. `https://product.hubspot.com/blog/git-and-github-tutorial-for-beginners`)
- Any YouTube video URL (e.g. `https://www.youtube.com/watch?v=0fKg7e37bQE`)

A complete `items` list might look like:

```
items:
 - pybee/batavia
 - Homebrew/brew
 - https://www.youtube.com/watch?v=dSl_qnWO104
```

### created_by
(if applicable) GitHub username of the person and/or organization that authored the collection.

### display_name
(required) The collection name that will be displayed on its page (ex. `How to choose (and contribute to) your first open source project`). Should use proper noun capitalization. Emoji are not allowed.

### The body of your document
(required) A longer description of the collection, which will appear on its page. Must be 1,000 characters or less. Some Markdown is allowed, such as links. Emoji are allowed.