Remove country topics from Hacktoberfest
As with the year topics, these topics are not tracked by Hacktoberfest and do not opt-in a repository. Having them listed may likely lead to contributors looking at repositories that are not opted-in for Hacktoberfest, and then having their contributions not be counted.

Further, all the topics linked here are essentially unused, returning less than 5 repositories each in their results.