Add 'script' and organize todo.
Added the 'script' topic. Could not find a good wikipedia article for it, so I just got the definition of a computer script.

I also organized the topics todo list alphabetically. Please correct if I mis-marked a topic, as I used a tool to sort it online, requiring me to mark everything as done.

Thanks,
Colack